address georges comments, change player on healthbar to targetdummy to test
