+++ Danki2/Assets/Scripts/UI/Bar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fcac7b1c25877c648b1d09363b3dedd7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
