+++ Danki2/Assets/Scripts/UI/Bars.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d65cd3c1f7cc8b41989d4fe0f30b9ce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
