+++ Danki2/Assets/Scripts/UI/Bars/ChannelBar.cs
@@ -2,13 +2,13 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class ChannelBar : MonoBehaviour
+public class ChannelBar : Bar
 {
     Player _player;
 
     void Awake()
     {
-        transform.localScale = new Vector3(0f, 1f, 1f);
+        SetWidth(0f);
         _player = FindObjectOfType<Player>();
     }
 
@@ -19,12 +19,10 @@ public class ChannelBar : MonoBehaviour
             && _player.CastingStatus != CastingStatus.ChannelingRight
         )
         {
-            transform.localScale = new Vector3(0f, 1f, 1f);
+            SetWidth(0f);
             return;
         }
 
-        var totalDuration = _player.ChannelService.TotalDuration;
-        var remainingDuration = _player.ChannelService.RemainingDuration;
-        transform.localScale = new Vector3(remainingDuration / totalDuration, 1f, 1f);
+        SetWidth(_player.ChannelService.RemainingDuration / _player.ChannelService.TotalDuration);
     }
 }
