+++ Danki2/Assets/Scripts/UI/Bars/HealthBar.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class HealthBar : Bar
+{
+    Player _player;
+
+    void Awake()
+    {
+        _player = FindObjectOfType<Player>();
+    }
+
+    void Update()
+    {
+        SetWidth(Mathf.Max(_player.Health, 0f) / _player.GetStat(Stat.MaxHealth));
+    }
+}
